cs444_genai_class_flow/src/cs444_genai_class_flow/crews/crew2/pyproject.toml
[project]
name = "crew2"
version = "0.1.0"
description = "crew2 using CrewAI to generate story, dialogue, and game audio"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"

dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0",
    "beautifulsoup4",
    "requests",
    "json5",
    "openai",
    "retrying",
    "python-dotenv"
]

[project.scripts]
crew2 = "crew2.main:run"
run_crew = "crew2.main:run"
train = "crew2.main:train"
replay = "crew2.main:replay"
test = "crew2.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
requirements = "requirements.txt"
